// src/app/modules/hcm/shared/models/holidays/holidays-balance.ts
export class HolidaysBalance{
    holidaysTotal: number = -1;
    accumulatedCyclePrevious: number = -1;
    programmed: number = -1;
    noProgrammed: number = -1;
    expired: number = -1;
    enjoyed: number = -1;
    additionalDays: number = -1;
    dateEntry: string = '';
    firstNames: string = ''; 
    firstSurname: string = '';
    idClassPayroll: number = -1;
    classPayroll: string = '';
    numberworker: string= '';
    idLaborRelationship: number = -1; 
}